requirements.txt
# Requirements for LeafEngines Agricultural Intelligence Skill
# These packages are needed if running Python scripts independently

requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
PyYAML>=6.0

# Optional for advanced features
# scikit-learn>=1.3.0  # For machine learning predictions
# matplotlib>=3.7.0    # For data visualization
# scipy>=1.11.0        # For scientific calculations